+++ DevAccelerationSystem/Assets/DevAccelerationSystem/Tests/Editor/BuildTargetExtensionTests.cs
@@ -1,11 +1,8 @@
 using System;
-using System.Collections;
-using System.Reflection;
 using DevAccelerationSystem.ProjectCompilationCheck;
 using NUnit.Framework;
 using UnityEditor;
 using UnityEngine;
-using UnityEngine.TestTools;
 
 namespace DevAccelerationSystem.Tests.Editor
 {
@@ -21,20 +18,7 @@ namespace DevAccelerationSystem.Tests.Editor
            //act
            var isSupported = act.Invoke();
            //assert
-           Assert.IsTrue(isSupported, $"The reflection in the version of Unity {Application.unityVersion} is broken. The method IsPlatformSupportLoadedByBuildTarget is not available");
+           Assert.IsTrue(isSupported, $"The active build target must be available in Unity {Application.unityVersion}.");
        }
-       
-        [Test]
-        public void IsPlatformSupportLoadedByBuildTarget_GetMethodViaReflection_IsAvailable()
-        {
-            //arrange
-            MethodInfo methodInfo = null;
-            Action act = () => methodInfo = BuildTargetExtension.IsPlatformSupportLoadedByBuildTargetGetMethod();
-            //act
-            act.Invoke();
-            //assert
-            Assert.IsNotNull(methodInfo, $"Could not get the method in the Unity version. Reflection of extracting has to be updated in the Unity verion {Application.unityVersion}");
-        }
-
     }
 }